"add 2 function log "
